+++ csrc/trans_config.h
@@ -62,10 +62,10 @@ public:
 
 class TransferTask {
 public:
-    TransferTask(const TransferTaskMeta& meta, int video_addr, TaskType type)
+    TransferTask(const TransferTaskMeta& meta, long long video_addr, TaskType type)
         : meta(meta), video_addr(video_addr), type(type){}
     TransferTaskMeta meta;
-    int video_addr;
+    long long video_addr;
     TaskType type;
 
     // Serialize the TransferTask to a string (JSON format)
@@ -81,7 +81,7 @@ public:
     static TransferTask deserialize(const std::string& serialized_data) {
         json task = json::parse(serialized_data);
         TransferTaskMeta meta = TransferTaskMeta::from_json(task.at("meta"));
-        int video_addr = task.at("video_addr").get<int>();
+        long long video_addr = task.at("video_addr").get<long long>();
         TaskType type = static_cast<TaskType>(task.at("type").get<int>());
         return TransferTask(meta, video_addr, type);
     }
